refactor: prune legacy iOS surfaces
