--- apps/ios/LogYourBody/Views/LogWeightView.swift
@@ -1,473 +1,0 @@
-//
-// LogWeightView.swift
-// LogYourBody
-//
-import SwiftUI
-
-struct LogWeightView: View {
-    @EnvironmentObject var authManager: AuthManager
-    @State private var weight: String = ""
-    @State private var bodyFat: String = ""
-    @State private var weightError: String?
-    @State private var bodyFatError: String?
-    @AppStorage(Constants.preferredMeasurementSystemKey) private var measurementSystem = PreferencesView.defaultMeasurementSystem
-    @State private var isLoading = false
-    @State private var showError = false
-    @State private var errorMessage = ""
-    @State private var showSuccess = false
-    @Environment(\.dismiss)
-    var dismiss
-    @FocusState private var focusedField: Field?
-
-    enum Field {
-        case weight, bodyFat
-    }
-
-    var currentSystem: MeasurementSystem {
-        MeasurementSystem.fromStored(rawValue: measurementSystem)
-    }
-
-    var body: some View {
-        NavigationStack {
-            ZStack {
-                // Background
-                Color.appBackground
-                    .ignoresSafeArea()
-
-                VStack(spacing: 0) {
-                    // Header
-                    HStack {
-                        Button(
-                            action: { dismiss() },
-                            label: {
-                                Image(systemName: "xmark")
-                                    .font(.system(size: 18, weight: .medium))
-                                    .foregroundColor(.appTextSecondary)
-                                    .frame(width: 44, height: 44)
-                            }
-                        )
-
-                        Spacer()
-
-                        Text("Log Metrics")
-                            .font(.system(size: 17, weight: .semibold))
-                            .foregroundColor(.appText)
-
-                        Spacer()
-
-                        // Quick add button (optional)
-                        Button(action: quickAdd) {
-                            Text("Quick")
-                                .font(.system(size: 15, weight: .medium))
-                                .foregroundColor(.appPrimary)
-                        }
-                        .frame(width: 44, height: 44)
-                    }
-                    .padding(.horizontal, 16)
-                    .padding(.top, 8)
-
-                    ScrollView {
-                        VStack(spacing: 32) {
-                            // Hero section with both inputs
-                            VStack(spacing: 24) {
-                                // Weight input card
-                                VStack(spacing: 16) {
-                                    HStack {
-                                        Image(systemName: "scalemass.fill")
-                                            .font(.system(size: 20))
-                                            .foregroundColor(.blue)
-                                        Text("Weight")
-                                            .font(.system(size: 18, weight: .medium))
-                                            .foregroundColor(.appText)
-                                        Spacer()
-                                    }
-
-                                    HStack(alignment: .bottom, spacing: 8) {
-                                        TextField("--", text: $weight)
-                                            .font(.system(size: 40, weight: .bold, design: .rounded))
-                                            .foregroundColor(.appText)
-                                            .multilineTextAlignment(.center)
-                                            .keyboardType(.decimalPad)
-                                            .focused($focusedField, equals: .weight)
-                                            .frame(maxWidth: 150)
-                                            .accessibilityIdentifier("log_weight_weight_field")
-                                            .onChange(of: weight) { _, newValue in
-                                                validateWeight(newValue)
-                                            }
-
-                                        Text(currentSystem.weightUnit)
-                                            .font(.system(size: 20, weight: .medium))
-                                            .foregroundColor(.appTextSecondary)
-                                            .padding(.bottom, 6)
-                                    }
-
-                                    if let weightError {
-                                        validationMessage(weightError)
-                                            .accessibilityIdentifier("log_weight_weight_error")
-                                    }
-                                }
-                                .padding(20)
-                                .background(Color.appCard)
-                                .cornerRadius(16)
-                                .overlay(
-                                    RoundedRectangle(cornerRadius: 16)
-                                        .stroke(
-                                            focusedField == .weight ? Color.appPrimary : Color.appBorder,
-                                            lineWidth: focusedField == .weight ? 2 : 1
-                                        )
-                                )
-                                .onTapGesture {
-                                    focusedField = .weight
-                                }
-
-                                // Body fat input card
-                                VStack(spacing: 16) {
-                                    HStack {
-                                        Image(systemName: "percent")
-                                            .font(.system(size: 20))
-                                            .foregroundColor(.orange)
-                                        Text("Body Fat")
-                                            .font(.system(size: 18, weight: .medium))
-                                            .foregroundColor(.appText)
-                                        Spacer()
-                                        Text("Optional")
-                                            .font(.system(size: 12, weight: .medium))
-                                            .foregroundColor(.appTextTertiary)
-                                            .padding(.horizontal, 8)
-                                            .padding(.vertical, 4)
-                                            .background(Color.appTextTertiary.opacity(0.1))
-                                            .cornerRadius(8)
-                                    }
-
-                                    HStack(alignment: .bottom, spacing: 8) {
-                                        TextField("--", text: $bodyFat)
-                                            .font(.system(size: 40, weight: .bold, design: .rounded))
-                                            .foregroundColor(.appText)
-                                            .multilineTextAlignment(.center)
-                                            .keyboardType(.decimalPad)
-                                            .focused($focusedField, equals: .bodyFat)
-                                            .frame(maxWidth: 150)
-                                            .accessibilityIdentifier("log_weight_body_fat_field")
-                                            .onChange(of: bodyFat) { _, newValue in
-                                                validateBodyFat(newValue)
-                                            }
-
-                                        Text("%")
-                                            .font(.system(size: 20, weight: .medium))
-                                            .foregroundColor(.appTextSecondary)
-                                            .padding(.bottom, 6)
-                                    }
-
-                                    if let bodyFatError {
-                                        validationMessage(bodyFatError)
-                                            .accessibilityIdentifier("log_weight_body_fat_error")
-                                    }
-                                }
-                                .padding(20)
-                                .background(Color.appCard)
-                                .cornerRadius(16)
-                                .overlay(
-                                    RoundedRectangle(cornerRadius: 16)
-                                        .stroke(
-                                            focusedField == .bodyFat ? Color.appPrimary : Color.appBorder,
-                                            lineWidth: focusedField == .bodyFat ? 2 : 1
-                                        )
-                                )
-                                .onTapGesture {
-                                    focusedField = .bodyFat
-                                }
-                            }
-                            .padding(.horizontal, 20)
-                            .padding(.top, 20)
-
-                            // Date and helpful info
-                            VStack(spacing: 12) {
-                                HStack {
-                                    Image(systemName: "calendar")
-                                        .font(.system(size: 14))
-                                        .foregroundColor(.appTextSecondary)
-                                    Text("Today, \(Date(), style: .date)")
-                                        .font(.system(size: 14))
-                                        .foregroundColor(.appTextSecondary)
-                                    Spacer()
-                                }
-
-                                if !weight.isEmpty || !bodyFat.isEmpty {
-                                    VStack(spacing: 8) {
-                                        if !weight.isEmpty, let weightValue = Double(weight) {
-                                            HStack {
-                                                Text("Weight:")
-                                                    .foregroundColor(.appTextSecondary)
-                                                Spacer()
-                                                Text("\(weightValue, specifier: "%.1f") \(currentSystem.weightUnit)")
-                                                    .foregroundColor(.appText)
-                                            }
-                                            .font(.system(size: 14))
-                                        }
-
-                                        if !bodyFat.isEmpty, let bfValue = Double(bodyFat) {
-                                            HStack {
-                                                Text("Body Fat:")
-                                                    .foregroundColor(.appTextSecondary)
-                                                Spacer()
-                                                Text("\(bfValue, specifier: "%.1f")%")
-                                                    .foregroundColor(.appText)
-                                            }
-                                            .font(.system(size: 14))
-                                        }
-                                    }
-                                    .padding(12)
-                                    .background(Color.appPrimary.opacity(0.05))
-                                    .cornerRadius(12)
-                                    .transition(.opacity.combined(with: .scale(scale: 0.95)))
-                                }
-                            }
-                            .padding(.horizontal, 20)
-                        }
-                        .padding(.bottom, 28)
-                    }
-
-                    // Save button (fixed at bottom)
-                    VStack(spacing: 12) {
-                        if showSuccess {
-                            HStack {
-                                Image(systemName: "checkmark.circle.fill")
-                                    .foregroundColor(.green)
-                                Text("Saved successfully!")
-                                    .font(.system(size: 14, weight: .medium))
-                                    .foregroundColor(.green)
-                            }
-                            .transition(.opacity.combined(with: .move(edge: .bottom)))
-                        }
-
-                        Button(action: saveMeasurement) {
-                            Group {
-                                if isLoading {
-                                    HStack(spacing: 12) {
-                                        ProgressView()
-                                            .progressViewStyle(CircularProgressViewStyle(tint: .white))
-                                            .scaleEffect(0.8)
-                                        Text("Saving...")
-                                            .font(.system(size: 17, weight: .semibold))
-                                            .foregroundColor(.white)
-                                    }
-                                } else {
-                                    Text("Save Measurement")
-                                        .font(.system(size: 17, weight: .semibold))
-                                        .foregroundColor(.white)
-                                }
-                            }
-                            .frame(height: 54)
-                            .frame(maxWidth: .infinity)
-                            .background(isFormValid() ? Color.appPrimary : Color.appBorder)
-                            .cornerRadius(16)
-                            .scaleEffect(isFormValid() ? 1.0 : 0.98)
-                            .animation(.spring(response: 0.3, dampingFraction: 0.7), value: isFormValid())
-                        }
-                        .disabled(!isFormValid() || isLoading)
-                        .accessibilityIdentifier("log_weight_save_measurement_button")
-                        .padding(.horizontal, 20)
-                        .padding(.bottom, 34)
-                    }
-                    .background(
-                        Color.appBackground
-                            .shadow(color: .black.opacity(0.05), radius: 20, x: 0, y: -10)
-                    )
-                }
-            }
-            .preferredColorScheme(.dark)
-            .onTapGesture {
-                hideKeyboard()
-            }
-            .standardErrorAlert(isPresented: $showError, message: errorMessage)
-        }
-        .onAppear {
-            // Auto-focus on weight field when view appears
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                focusedField = .weight
-            }
-        }
-    }
-
-    private func hideKeyboard() {
-        UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
-    }
-
-    private func isFormValid() -> Bool {
-        LogWeightFormValidator
-            .validate(weight: weight, bodyFat: bodyFat, unit: currentSystem.weightUnit)
-            .isValid
-    }
-
-    private func validateWeight(_ value: String) {
-        weightError = LogWeightFormValidator.fieldError(for: value, field: .weight, unit: currentSystem.weightUnit)
-    }
-
-    private func validateBodyFat(_ value: String) {
-        bodyFatError = LogWeightFormValidator.fieldError(for: value, field: .bodyFat, unit: currentSystem.weightUnit)
-    }
-
-    private func validationMessage(_ message: String) -> some View {
-        HStack(spacing: 6) {
-            Image(systemName: "exclamationmark.circle.fill")
-                .font(.system(size: 12, weight: .semibold))
-            Text(message)
-                .font(.system(size: 13, weight: .medium))
-                .fixedSize(horizontal: false, vertical: true)
-            Spacer(minLength: 0)
-        }
-        .foregroundColor(.red)
-    }
-
-    private func quickAdd() {
-        Task {
-            // Get last logged values for quick entry
-            guard let userId = authManager.currentUser?.id else { return }
-            let fromDate = Calendar.current.date(byAdding: .day, value: -30, to: Date()) ?? Date()
-            let cachedMetrics = await CoreDataManager.shared.fetchBodyMetrics(for: userId, from: fromDate, to: Date())
-            let recentMetrics = cachedMetrics.compactMap { $0.toBodyMetrics() }.sorted { $0.date > $1.date }
-
-            if let latest = recentMetrics.first {
-                await MainActor.run {
-                    if weight.isEmpty, let latestWeight = latest.weight {
-                        let displayWeight = currentSystem.weightUnit == "lbs" ? latestWeight.kgToLbs : latestWeight
-                        weight = String(format: "%.1f", displayWeight)
-                    }
-                    if bodyFat.isEmpty, let latestBodyFat = latest.bodyFatPercentage {
-                        bodyFat = String(format: "%.1f", latestBodyFat)
-                    }
-                }
-            }
-        }
-    }
-
-    private func saveMeasurement() {
-        let validation = LogWeightFormValidator.validate(weight: weight, bodyFat: bodyFat, unit: currentSystem.weightUnit)
-        weightError = validation.weightError
-        bodyFatError = validation.bodyFatError
-
-        guard validation.isValid else {
-            if let formError = validation.formError {
-                errorMessage = formError
-                showError = true
-            }
-            return
-        }
-
-        isLoading = true
-        hideKeyboard()
-
-        Task {
-            do {
-                _ = try await BodyMetricLoggingService.shared.log(
-                    weight: weight,
-                    bodyFat: bodyFat,
-                    unit: currentSystem.weightUnit
-                )
-
-                await MainActor.run {
-                    isLoading = false
-                    showSuccess = true
-                }
-
-                // Show success briefly then dismiss
-                try await Task.sleep(nanoseconds: 1_000_000_000) // 1 second
-
-                await MainActor.run {
-                    dismiss()
-                }
-            } catch {
-                await MainActor.run {
-                    isLoading = false
-                    errorMessage = error.localizedDescription
-                    showError = true
-                }
-            }
-        }
-    }
-}
-
-enum LogWeightInputField {
-    case weight
-    case bodyFat
-}
-
-struct LogWeightFormValidation: Equatable {
-    let weightValue: Double?
-    let bodyFatValue: Double?
-    let weightError: String?
-    let bodyFatError: String?
-    let formError: String?
-
-    var isValid: Bool {
-        formError == nil
-            && weightError == nil
-            && bodyFatError == nil
-            && (weightValue != nil || bodyFatValue != nil)
-    }
-}
-
-enum LogWeightFormValidator {
-    static func validate(weight: String, bodyFat: String, unit: String) -> LogWeightFormValidation {
-        var weightValue: Double?
-        var bodyFatValue: Double?
-        var weightError: String?
-        var bodyFatError: String?
-
-        if hasValue(weight) {
-            do {
-                weightValue = try ValidationService.shared.validateWeight(weight, unit: unit)
-            } catch let error as ValidationError {
-                weightError = error.errorDescription
-            } catch {
-                weightError = "Please enter a valid number"
-            }
-        }
-
-        if hasValue(bodyFat) {
-            do {
-                bodyFatValue = try ValidationService.shared.validateBodyFat(bodyFat)
-            } catch let error as ValidationError {
-                bodyFatError = error.errorDescription
-            } catch {
-                bodyFatError = "Please enter a valid percentage"
-            }
-        }
-
-        let formError = !hasValue(weight) && !hasValue(bodyFat)
-            ? "Please enter at least one measurement"
-            : nil
-
-        return LogWeightFormValidation(
-            weightValue: weightValue,
-            bodyFatValue: bodyFatValue,
-            weightError: weightError,
-            bodyFatError: bodyFatError,
-            formError: formError
-        )
-    }
-
-    static func fieldError(for value: String, field: LogWeightInputField, unit: String) -> String? {
-        guard hasValue(value) else { return nil }
-
-        switch field {
-        case .weight:
-            return validate(weight: value, bodyFat: "", unit: unit).weightError
-        case .bodyFat:
-            return validate(weight: "", bodyFat: value, unit: unit).bodyFatError
-        }
-    }
-
-    private static func hasValue(_ value: String) -> Bool {
-        !value.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-}
-
-struct LogWeightView_Previews: PreviewProvider {
-    static var previews: some View {
-        LogWeightView()
-            .environmentObject(AuthManager.shared)
-            .environmentObject(RealtimeSyncManager.shared)
-            .preferredColorScheme(.dark)
-    }
-}
